requirements.txt
numpy
pandas==2.2.2
torch

einops==0.8.1
huggingface_hub==0.33.1
matplotlib==3.9.3
tqdm==4.67.1
safetensors==0.4.3
python-binance
cryptography
aiohttp
urllib3==2.2.2
requests
tenacity==8.2.3
vercel-blob==0.4.2
python-dotenv==1.0.1
akshare
pytz
yfinance